services/mission-compiler/requirements.txt
pydantic>=2.0
pyyaml>=6.0
jsonschema>=4.0
